+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## vNext
+* Add webpack alias to consume the component library as ES2015 modules in production
+
 ## v0.6.0-rc.2 (2017-07-11) **[pre-release @next]**
 
 * Fix TypeScript module setting override for TypeScript < 2.4